+++ app/files/changelog.md
@@ -1,5 +1,5 @@
 ## Version Pf9.0.0 - February 2026
-- Ingested the MalariaGEN Pf9 data release - now 40,074 QC-passed samples. Alongside samples sequenced at the Sanger Institute, Pf9 now includes ENA samples sequenced by other organisations. Some samples were sequenced because particular kelch13 mutations were detected. Haplotype frequencies for these locations and years are therefore likely to be overestimated and should be treated with caution: Uganda (2017–2022) selected for C469Y or A675V; Kagera, Tanzania (2021–2022) selected for R561H; Guyana (2016) selected for C580Y.
+- Ingested the MalariaGEN Pf9 data release - now 40,074 QC-passed samples. Alongside samples sequenced at the Sanger Institute, Pf9 now includes ENA samples sequenced by other organisations. A small proportion of samples were selected because they had particular kelch13 mutations initially detected by malaria molecular surveillance. Haplotype frequencies for the following locations are therefore likely overestimated and should be treated with caution: Uganda (2017-2022) selected for C469Y or A675V; Kagera, Tanzania (2021 and 2022) selected for R561H; Guyana (2016) selected for C580Y.
 - Previous Pf-HaploAtlas Pf8 version has been made available [here](https://pf8-haploatlas.streamlit.app/) (it may require waking up the app!).